layout/generic/test/test_bug384527.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=384527
-->
<head style="display: inline-table;">
  <title>Test for Bug 384527</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <style style="display: block; direction: rtl;">
   style::first-letter {float: right;}
  </style>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=384527">Mozilla Bug 384527</a>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
 function do_Test()
{
/** Test for Bug 384527 **/
  ok(true, "Should not crash");
  SimpleTest.finish();
}
  SimpleTest.requestFlakyTimeout("untriaged");
  setTimeout(do_Test, 500);
  SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
